+++ packages/video_player/video_player/pigeons/messages.dart
@@ -14,6 +14,11 @@ class VolumeMessage {
   double volume;
 }
 
+class PlaybackSpeedMessage {
+  int textureId;
+  double speed;
+}
+
 class PositionMessage {
   int textureId;
   int position;
@@ -37,6 +42,7 @@ abstract class VideoPlayerApi {
   void dispose(TextureMessage msg);
   void setLooping(LoopingMessage msg);
   void setVolume(VolumeMessage msg);
+  void setPlaybackSpeed(PlaybackSpeedMessage msg);
   void play(TextureMessage msg);
   PositionMessage position(TextureMessage msg);
   void seekTo(PositionMessage msg);