+++ packages/video_player/video_player_platform_interface/test/method_channel_video_player_test.dart
@@ -4,13 +4,12 @@
 
 import 'dart:ui';
 
-import 'package:mockito/mockito.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-
+import 'package:mockito/mockito.dart';
+import 'package:video_player_platform_interface/messages.dart';
 import 'package:video_player_platform_interface/method_channel_video_player.dart';
 import 'package:video_player_platform_interface/video_player_platform_interface.dart';
-import 'package:video_player_platform_interface/messages.dart';
 
 class _ApiLogger implements VideoPlayerApiTest {
   final List<String> log = [];
@@ -19,6 +18,7 @@ class _ApiLogger implements VideoPlayerApiTest {
   PositionMessage positionMessage;
   LoopingMessage loopingMessage;
   VolumeMessage volumeMessage;
+  PlaybackSpeedMessage playbackSpeedMessage;
   MixWithOthersMessage mixWithOthersMessage;
 
   @override
@@ -80,6 +80,12 @@ class _ApiLogger implements VideoPlayerApiTest {
   void setVolume(VolumeMessage arg) {
     log.add('setVolume');
     volumeMessage = arg;
+  }
+
+  @override
+  void setPlaybackSpeed(PlaybackSpeedMessage arg) {
+    log.add('setPlaybackSpeed');
+    playbackSpeedMessage = arg;
   }
 }
 
@@ -201,6 +207,13 @@ void main() {
       expect(log.log.last, 'setVolume');
       expect(log.volumeMessage.textureId, 1);
       expect(log.volumeMessage.volume, 0.7);
+    });
+
+    test('setPlaybackSpeed', () async {
+      await player.setPlaybackSpeed(1, 1.5);
+      expect(log.log.last, 'setPlaybackSpeed');
+      expect(log.playbackSpeedMessage.textureId, 1);
+      expect(log.playbackSpeedMessage.speed, 1.5);
     });
 
     test('seekTo', () async {